+++ examples/increment.spec.ts
@@ -1,5 +1,5 @@
 import { IncrementContract } from "./output/increment_IncrementContract";
-import { ContractSystem } from "ton-emulator";
+import { ContractSystem } from "@tact-lang/emulator";
 import { toNano } from "ton-core";
 
 describe('increment', () => {