+++ src/test/feature-debug.spec.ts
@@ -1,5 +1,5 @@
 import { toNano } from 'ton-core';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem, Verbosity } from '@tact-lang/emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { Debug } from './features/output/debug_Debug';