+++ src/test/feature-instrinsics.spec.ts
@@ -1,5 +1,5 @@
 import { Address, beginCell, Cell, toNano } from 'ton-core';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem } from '@tact-lang/emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { IntrinsicsTester } from './features/output/intrinsics_IntrinsicsTester';