+++ src/test/feature-map.spec.ts
@@ -1,7 +1,7 @@
 import { randomAddress } from './utils/randomAddress';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { MapTestContract, SomeStruct } from './features/output/maps_MapTestContract';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem } from '@tact-lang/emulator';
 import { beginCell, toNano } from 'ton-core';
 import { ComputeError } from 'ton-core';