+++ src/test/feature-masterchain.spec.ts
@@ -1,5 +1,5 @@
 import { Address, beginCell, toNano } from 'ton-core';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem } from '@tact-lang/emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { MasterchainTester } from './features/output/masterchain_MasterchainTester';
 import { MasterchainTester as EnabledTester } from './features/output/masterchain-allow_MasterchainTester';