+++ src/test/feature-ordering.spec.ts
@@ -1,5 +1,5 @@
 import { toNano } from 'ton-core';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem } from '@tact-lang/emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { A } from './features/output/ordering_A';