+++ src/test/feature-random.spec.ts
@@ -1,5 +1,5 @@
 import { toNano } from 'ton-core';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem } from '@tact-lang/emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { RandomContract } from './features/output/random_RandomContract';