+++ src/test/feature-strings.spec.ts
@@ -1,5 +1,5 @@
 import { beginCell, toNano } from 'ton-core';
-import { ContractSystem } from 'ton-emulator';
+import { ContractSystem } from '@tact-lang/emulator';
 import { __DANGER_resetNodeId } from '../grammar/ast';
 import { StringsTester } from './features/output/strings_StringsTester';